+++ IfcPlugins/src/org/bimserver/ifc/step/deserializer/XPass.java
@@ -27,7 +27,7 @@ import com.google.common.base.Charsets;
 public class XPass extends Pass {
 
 	@Override
-	public String process(int lineNumber, String result) throws DeserializeException {
+	public String process(long lineNumber, String result) throws DeserializeException {
 		while (result.contains("\\X\\")) {
 			int index = result.indexOf("\\X\\");
 			int code = Integer.parseInt(result.substring(index + 3, index + 5), 16);
