+++ IfcPlugins/src/org/bimserver/ifc/step/serializer/IfcStepSerializer.java
@@ -172,7 +172,7 @@ public class IfcStepSerializer extends IfcSerializer implements HeaderTakingSeri
 			return;
 		}
 		print(DASH);
-		int convertedKey = getExpressId(object);
+		long convertedKey = getExpressId(object);
 		if (convertedKey == -1) {
 			throw new SerializerException("Going to serialize an object with id -1 (" + object.eClass().getName() + ")");
 		}
